add implementation of chap 10,11,12

+++ runtime.c
@@ -6,7 +6,7 @@
 
 int *initArray(int size, int init)
 {int i;
- int *a = (int *)malloc(size*sizeof(int));
+ int *a = (int *)malloc(size*sizeof(long));
  for(i=0;i<size;i++) a[i]=init;
  return a;
 }
